requirements.txt
# Mitch Discord Bot - Python Dependencies
# v1.0.0 - Production-Ready with AI Integration

# Discord bot framework (includes aiohttp)
discord.py==2.3.2

# YAML configuration parsing
PyYAML==6.0.1

# HTTP client for Ollama API (included with discord.py, listed for clarity)
# aiohttp>=3.8.0

# Testing framework
pytest==7.4.3
